backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Data Validation
pydantic==2.10.5
pydantic-settings==2.7.1

# Authentication & Security
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML Libraries
groq==0.11.0
sentence-transformers==3.3.1
pinecone-client==5.0.1
tiktoken==0.8.0

# Document Processing
pymupdf==1.26.7
python-docx==1.1.2

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Caching
redis==5.2.1

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
tenacity==9.0.0

# Scientific Computing
numpy==2.2.1
scikit-learn==1.6.1
